# extracted_project/Session-Guard/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Telegram Broadcast Bot API
 * OpenAPI spec version: 0.1.0
 */
export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface MobileTokenExchangeRequest {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  code_verifier: string;
  /** @minLength 1 */
  redirect_uri: string;
  /** @minLength 1 */
  state: string;
  /** @minLength 1 */
  nonce?: string;
}

export interface MobileTokenExchangeSuccess {
  token: string;
}

export const LogoutSuccessValue = {
  success: true,
} as const;
export type LogoutSuccess = typeof LogoutSuccessValue;

export interface ErrorEnvelope {
  error: string;
}

export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface DashboardStats {
  totalBots: number;
  totalSubscribers: number;
  pendingRequests: number;
  totalTokens: number;
  activeTokens: number;
  totalBroadcasts: number;
}

export interface Bot {
  id: number;
  name: string;
  token: string;
  /** @nullable */
  username?: string | null;
  /** @nullable */
  superuserChatId?: string | null;
  /** @nullable */
  paymentInfo?: string | null;
  isActive: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface CreateBotRequest {
  name: string;
  token: string;
}

export interface UpdateBotRequest {
  name?: string;
  superuserChatId?: string;
  paymentInfo?: string;
}

export interface WebhookSetupResponse {
  webhookUrl: string;
}

export interface Subscriber {
  id: number;
  botId: number;
  chatId: string;
  /** @nullable */
  username?: string | null;
  /** @nullable */
  firstName?: string | null;
  /** @nullable */
  lastName?: string | null;
  isActive: boolean;
  subscribedAt: string;
}

export interface AddSubscriberRequest {
  chatId: string;
  username?: string;
  firstName?: string;
  lastName?: string;
}

export type BroadcastStatus =
  (typeof BroadcastStatus)[keyof typeof BroadcastStatus];

export const BroadcastStatus = {
  draft: "draft",
  sending: "sending",
  completed: "completed",
  failed: "failed",
} as const;

export interface Broadcast {
  id: number;
  botId: number;
  message: string;
  status: BroadcastStatus;
  totalRecipients: number;
  sentCount: number;
  failedCount: number;
  createdAt: string;
  /** @nullable */
  completedAt?: string | null;
}

export interface CreateBroadcastRequest {
  message: string;
}

export interface Plan {
  id: number;
  botId: number;
  name: string;
  durationMonths: number;
  price: string;
  isActive: boolean;
  createdAt: string;
}

export interface CreatePlanRequest {
  name: string;
  durationMonths: number;
  price: string;
}

export type PaymentRequestStatus =
  (typeof PaymentRequestStatus)[keyof typeof PaymentRequestStatus];

export const PaymentRequestStatus = {
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
} as const;

export interface PaymentRequest {
  id: number;
  botId: number;
  chatId: string;
  /** @nullable */
  telegramUsername?: string | null;
  /** @nullable */
  telegramFirstName?: string | null;
  planId: number;
  status: PaymentRequestStatus;
  /** @nullable */
  rejectionReason?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface Token {
  id: number;
  token: string;
  botId: number;
  planId: number;
  /** @nullable */
  paymentRequestId?: number | null;
  isActivated: boolean;
  /** @nullable */
  activatedByChatId?: string | null;
  /** @nullable */
  activatedAt?: string | null;
  expiresAt: string;
  createdAt: string;
}

export interface ApprovePaymentResponse {
  token: Token;
}

export interface RejectPaymentRequest {
  reason?: string;
}

export interface Userbot {
  id: number;
  botId: number;
  chatId: string;
  phone: string;
  /** @nullable */
  telegramUserId?: string | null;
  /** @nullable */
  telegramUsername?: string | null;
  /** @nullable */
  telegramFirstName?: string | null;
  isActive: boolean;
  createdAt: string;
  updatedAt: string;
}

export type AuthorizationSessionHeaderParameter = string;

export type BeginBrowserLoginParams = {
  returnTo?: string;
};

export type HandleBrowserLoginCallbackParams = {
  code?: string;
  state?: string;
  iss?: string;
};