# extracted_project/Session-Guard/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Telegram Broadcast Bot API
servers:
  - url: /api
    description: Base API path
...
paths:
  /auth/user:
    get:
      tags: [Auth]
      operationId: getCurrentAuthUser
      summary: Get the currently authenticated user
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '200':
          description: Auth status resolved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUserEnvelope'
  /login:
    get:
      tags: [Auth]
      operationId: beginBrowserLogin
      summary: Start the browser OIDC login flow
      parameters:
        - name: returnTo
          in: query
          required: false
          schema:
            type: string
      responses:
        '302':
          description: Redirect to the OIDC authorization endpoint.
  /callback:
    get:
      tags: [Auth]
      operationId: handleBrowserLoginCallback
      summary: Complete the browser OIDC login flow
      parameters:
        - name: code
          in: query
          required: false
          schema:
            type: string
        - name: state
          in: query
          required: false
          schema:
            type: string
        - name: iss
          in: query
          required: false
          schema:
            type: string
            format: uri
      responses:
        '302':
          description: Redirect to the returnTo path on success or /api/login on failure.
  /logout:
    get:
      tags: [Auth]
      operationId: logoutBrowserSession
      summary: Clear the session and begin OIDC logout
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '302':
          description: Redirect to the OIDC provider logout URL.
  /mobile-auth/token-exchange:
    post:
      tags: [Auth]
      operationId: exchangeMobileAuthorizationCode
      summary: Exchange a mobile OIDC code for a session token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MobileTokenExchangeRequest'
      responses:
        '200':
          description: Session created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MobileTokenExchangeSuccess'
        '400':
          description: Missing or invalid required fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Token did not contain usable claims.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Token exchange failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /mobile-auth/logout:
    post:
      tags: [Auth]
      operationId: logoutMobileSession
      summary: Delete a mobile session token
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
      responses:
        '200':
          description: Logout processed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogoutSuccess'
...
  /bots/{botId}/broadcasts:
    parameters:
      - $ref: "#/components/parameters/BotId"
    get:
      operationId: listBroadcasts
      tags: [broadcasts]
      summary: List broadcasts for a bot
      responses:
        "200":
          description: Broadcast list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Broadcast"
    post:
      operationId: createBroadcast
      tags: [broadcasts]
      summary: Create and send a broadcast
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBroadcastRequest"
      responses:
        "201":
          description: Broadcast queued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Broadcast"
  /bots/{botId}/broadcasts/{broadcastId}:
    parameters:
      - $ref: "#/components/parameters/BotId"
      - $ref: "#/components/parameters/BroadcastId"
    get:
      operationId: getBroadcast
      tags: [broadcasts]
      summary: Get broadcast details
      responses:
        "200":
          description: Broadcast details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Broadcast"
  /bots/{botId}/plans:
    parameters:
      - $ref: "#/components/parameters/BotId"
    get:
      operationId: listPlans
      tags: [plans]
      summary: List plans for a bot
      responses:
        "200":
          description: Plan list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Plan"
    post:
      operationId: createPlan
      tags: [plans]
      summary: Create a subscription plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePlanRequest"
      responses:
        "201":
          description: Plan created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plan"
  /bots/{botId}/plans/{planId}:
    parameters:
      - $ref: "#/components/parameters/BotId"
      - $ref: "#/components/parameters/PlanId"
    delete:
      operationId: deletePlan
      tags: [plans]
      summary: Delete a plan
      responses:
        "204":
          description: Deleted
...
  /bots/{botId}/tokens:
    parameters:
      - $ref: "#/components/parameters/BotId"
    get:
      operationId: listTokens
      tags: [tokens]
      summary: List tokens for a bot
      responses:
        "200":
          description: Token list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Token"
  /bots/{botId}/userbots:
    parameters:
      - $ref: "#/components/parameters/BotId"
    get:
      operationId: listUserbots
      tags: [userbots]
      summary: List userbots (MTProto sessions) for a bot
      responses:
        "200":
          description: Userbot list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Userbot"
  /bots/{botId}/userbots/{chatId}:
    parameters:
      - $ref: "#/components/parameters/BotId"
      - name: chatId
        in: path
        required: true
        schema:
          type: string
    delete:
      operationId: revokeUserbot
      tags: [userbots]
      summary: Revoke a userbot session
      responses:
        "204":
          description: Revoked
components:
  parameters:
    AuthorizationSessionHeader:
      name: Authorization
      in: header
      required: false
      schema:
        type: string
        description: 'Bearer <session-token> for mobile clients'
    SessionCookie:
      name: sid
      in: cookie
      required: false
      schema:
        type: string
    BotId:
      name: botId
      in: path
      required: true
      schema:
        type: integer
...
    BroadcastId:
      name: broadcastId
      in: path
      required: true
      schema:
        type: integer
    PlanId:
      name: planId
      in: path
      required: true
      schema:
        type: integer
...
  schemas:
    AuthUser:
      type: object
      required: [id, email, firstName, lastName, profileImageUrl]
      properties:
        id:
          type: string
        email:
          type: [string, 'null']
          format: email
        firstName:
          type: [string, 'null']
        lastName:
          type: [string, 'null']
        profileImageUrl:
          type: [string, 'null']
    AuthUserEnvelope:
      type: object
      required: [user]
      properties:
        user:
          oneOf:
            - $ref: '#/components/schemas/AuthUser'
            - type: 'null'
    MobileTokenExchangeRequest:
      type: object
      required: [code, code_verifier, redirect_uri, state]
      properties:
        code:
          type: string
          minLength: 1
        code_verifier:
          type: string
          minLength: 1
        redirect_uri:
          type: string
          minLength: 1
          format: uri
        state:
          type: string
          minLength: 1
        nonce:
          type: string
          minLength: 1
    MobileTokenExchangeSuccess:
      type: object
      required: [token]
      properties:
        token:
          type: string
    LogoutSuccess:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          const: true
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Broadcast:
      type: object
      properties:
        id:
          type: integer
        botId:
          type: integer
        message:
          type: string
        status:
          type: string
          enum: [draft, sending, completed, failed]
        totalRecipients:
          type: integer
        sentCount:
          type: integer
        failedCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        completedAt:
          type: [string, 'null']
          format: date-time
      required: [id, botId, message, status, totalRecipients, sentCount, failedCount, createdAt]
    CreateBroadcastRequest:
      type: object
      properties:
        message:
          type: string
      required: [message]
    Plan:
      type: object
      properties:
        id:
          type: integer
        botId:
          type: integer
        name:
          type: string
        durationMonths:
          type: integer
        price:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, botId, name, durationMonths, price, isActive, createdAt]
    CreatePlanRequest:
      type: object
      properties:
        name:
          type: string
        durationMonths:
          type: integer
        price:
          type: string
      required: [name, durationMonths, price]
...
    Token:
      type: object
      properties:
        id:
          type: integer
        token:
          type: string
        botId:
          type: integer
        planId:
          type: integer
        paymentRequestId:
          type: [integer, 'null']
        isActivated:
          type: boolean
        activatedByChatId:
          type: [string, 'null']
        activatedAt:
          type: [string, 'null']
          format: date-time
        expiresAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
      required: [id, token, botId, planId, isActivated, expiresAt, createdAt]
    Userbot:
      type: object
      properties:
        id:
          type: integer
        botId:
          type: integer
        chatId:
          type: string
        phone:
          type: string
        telegramUserId:
          type: [string, 'null']
        telegramUsername:
          type: [string, 'null']
        telegramFirstName:
          type: [string, 'null']
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, botId, chatId, phone, isActive, createdAt, updatedAt]