# extracted_project/Session-Guard/lib/api-zod/src/generated/types/createPlanRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Telegram Broadcast Bot API
 * OpenAPI spec version: 0.1.0
 */

export interface CreatePlanRequest {
  name: string;
  durationMonths: number;
  price: string;
}